Separação de código com implementação de instruções em arquivos
