+++ include/bytecode/impl/_math.h
@@ -1,0 +1,12 @@
+#ifndef BYTECODE_IMPL_MATH_H
+#define BYTECODE_IMPL_MATH_H
+
+#include "interpreter.h"
+#include "Frame.h"
+
+void iadd(Frame *);
+void isub(Frame *);
+void imul(Frame *);
+void iinc(Frame *);
+
+#endif
