+++ include/bytecode/impl/comparisons.h
@@ -1,0 +1,10 @@
+#ifndef BYTECODE_IMPL_COMPARISONS_H
+#define BYTECODE_IMPL_COMPARISONS_H
+
+#include "Frame.h"
+#include "interpreter.h"
+
+void if_cond(Frame *);
+void if_icmp_cond(Frame *);
+
+#endif
