+++ include/bytecode/impl/constants.h
@@ -1,0 +1,15 @@
+#ifndef BYTECODE_IMPL_CONSTANTS_H
+#define BYTECODE_IMPL_CONSTANTS_H
+
+#include "Frame.h"
+#include "interpreter.h"
+
+void iconst_i(Frame *);
+void fconst_f(Frame *);
+
+void bipush(Frame *);
+void sipush(Frame *);
+
+void ldc_(Frame *);
+
+#endif
