+++ include/bytecode/impl/conversions.h
@@ -1,0 +1,9 @@
+#ifndef BYTECODE_IMPL_CONVERSIONS_H
+#define BYTECODE_IMPL_CONVERSIONS_H
+
+#include "Frame.h"
+#include "interpreter.h"
+
+void i2T(Frame *);
+
+#endif
