+++ include/bytecode/impl/stack.h
@@ -1,0 +1,9 @@
+#ifndef BYTECODE_IMPL_STACK_H
+#define BYTECODE_IMPL_STACK_H
+
+#include "Frame.h"
+
+void dup(Frame *);
+void dup2(Frame *);
+
+#endif
